Close #58 Team handlers extracted from user handlers file, the same with tests, user registration link creation moved to utils.py
