+++ backend/flaskr/model/utils.py
@@ -1,8 +1,8 @@
 from bson import ObjectId
 from flask import url_for
 
+from ..auth import as_jwt
 from ..email import send_email
-from backend.flaskr.auth import as_jwt
 
 
 def parse_id(data):
@@ -25,8 +25,8 @@ def check_question_type(question_type):
         raise ValueError(error_msg)
 
 
-def create_user_registration_link(user):
-    token = as_jwt({"username": user.name})
+def create_user_registration_link(username):
+    token = as_jwt({"username": username})
     return url_for('confirm', token=token, _external=True)
 
 
@@ -43,3 +43,23 @@ def invite_user_to_team(team, user):
                username=user.username,
                link=create_team_invitation_for_user_link(team, user.username),
                team_name=team.name)
+
+
+def confirm_user(link=None):
+    from ..database.unconfirmed_user_dao import UnconfirmedUserDAO
+    from ..database.user_dao import UserDAO
+    from ..model.user import User
+
+    if not link:
+        raise ValueError("No registration link provided")
+
+    unconfirmed_dao = UnconfirmedUserDAO()
+    query = {"link": link}
+    user = unconfirmed_dao.pop(query)
+
+    if user:
+        confirmed_user_data = user.user_data
+        confirmed_user = User(confirmed_user_data, password_hash=True)
+        confirmed_dao = UserDAO()
+        confirmed_dao.insert_one(confirmed_user)
+    return user
