+++ backend/flaskr/tests/test_team_dao.py
@@ -17,7 +17,7 @@ def clear_collection():
 def stub_team():
     data = {
         "name": "stub_team",
-        "users": []
+        "members": []
     }
     team = Team(data)
     return team
@@ -27,7 +27,7 @@ def stub_team():
 def stub_team_2():
     data = {
         "name": "stub_team_2",
-        "users": ["user_2"]
+        "members": ["user_2"]
     }
     team = Team(data)
     return team
@@ -51,7 +51,7 @@ def test_insert_one(clear_collection, stub_team):
 def test_nonexistent_user(clear_collection):
     error_data = {
         "name": "false_team",
-        "users": []
+        "members": []
     }
     nonexistent_team = Team(error_data)
