+++ backend/flaskr/tests/test_user_handler.py
@@ -5,12 +5,10 @@ from . import get_with_auth, post, post_with_auth, flask_client, \
     stub_user, clear_db, get, app
 from ..database import db
 from ..database.user_dao import UserDAO
-from ..database.unconfirmed_user_dao import UserDAO as UnconfirmedUserDAO
-from ..database.team_dao import TeamDAO
+from ..database.unconfirmed_user_dao import UnconfirmedUserDAO
 from ..auth import from_jwt
-from ..model.user import User
-from ..model.team import Team
 from ..model.unconfirmed_user import UnconfirmedUser
+from ..model.utils import create_user_registration_link
 
 
 # token acquire test
@@ -108,7 +106,7 @@ def test_confirm_user(clear_db, flask_client):
 
     with app.test_client(), app.test_request_context():
         data = {
-            "link": UnconfirmedUser.create_registration_link(user_data["username"]),
+            "link": create_user_registration_link(user_data["username"]),
             "user": user_data
         }
 
@@ -135,7 +133,7 @@ def test_confirmed_user_confirm(clear_db, flask_client):
 
     with app.test_client(), app.test_request_context():
         data = {
-            "link": UnconfirmedUser.create_registration_link(user_data["username"]),
+            "link": create_user_registration_link(user_data["username"]),
             "user": user_data
         }
 
@@ -151,35 +149,3 @@ def test_confirmed_user_confirm(clear_db, flask_client):
     res = get(flask_client, "/users/confirm/" + token)
     assert res.status_code == 200
     assert res.get_json()["confirmation"] == "Error"
-
-
-def test_confirm_invitation(clear_db, flask_client):
-    team_data = {
-        "name": "test_team",
-        "users": []
-    }
-    team = Team(team_data)
-
-    user_data = {
-        "username": "new_user",
-        "password": "123456789",
-        "email": "stubmail@gmail.com"
-    }
-    user = User(user_data)
-
-    user_dao = UserDAO()
-    team_dao = TeamDAO()
-
-    user_dao.insert_one(user)
-    team_dao.insert_one(team)
-
-    with app.test_client(), app.test_request_context():
-        invitation_link = team.get_invitation_link(user.username)
-
-    token = invitation_link.split('/')[-1]
-    res = get(flask_client, "/users/confirm_team/" + token)
-    assert res.status_code == 200
-    assert res.get_json()["confirmation"] == "OK"
-
-    team_from_db = team_dao.find_one_by_name(team.name)
-    assert user.username in team_from_db.members
